requirements.txt
streamlit
google-analytics-data
pandas